back/requirements.txt
asgiref==3.5.0
certifi==2023.7.22
charset-normalizer==3.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.12
django-ckeditor==6.4.0
django-cors-headers==3.14.0
django-filter==23.3
django-js-asset==2.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.6
Faker==18.6.2
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
Pillow==9.1.0
psycopg2==2.9.3
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==2.0.6
